+++ cmd/api/server.go
@@ -21,7 +21,7 @@ import (
 	"github.com/dolittle/platform-api/pkg/platform/job"
 	jobK8s "github.com/dolittle/platform-api/pkg/platform/job/k8s"
 	platformK8s "github.com/dolittle/platform-api/pkg/platform/k8s"
-	"github.com/dolittle/platform-api/pkg/platform/listeners"
+	m3ConnectorListeners "github.com/dolittle/platform-api/pkg/platform/listeners/m3connector"
 	"github.com/dolittle/platform-api/pkg/platform/microservice"
 	"github.com/dolittle/platform-api/pkg/platform/microservice/environmentVariables"
 	"github.com/dolittle/platform-api/pkg/platform/microservice/purchaseorderapi"
@@ -122,7 +122,7 @@ var serverCMD = &cobra.Command{
 		go job.NewCustomerJobListener(k8sClient, gitRepo, logContext.WithField("context", "listener-job-customer"))
 		go job.NewApplicationJobListener(k8sClient, gitRepo, logContext.WithField("context", "listener-job-application"))
 
-		go listeners.NewM3ConnectorConfigmapListener(k8sClient, gitRepo, gitRepo, logContext.WithField("context", "listener-m3connector-kafka-files"))
+		go m3ConnectorListeners.NewKafkaFilesConfigmapListener(k8sClient, gitRepo, gitRepo, logContext.WithField("context", "listener-m3connector-kafka-files"))
 
 		microserviceService := microservice.NewService(
 			isProduction,
